added a find time difference hook and integrated it
