+++ src/components/ledger/ComponentToPrint.js
@@ -1,24 +1,10 @@
 import React, { forwardRef, useImperativeHandle } from 'react';
 import { PDFDownloadLink } from '@react-pdf/renderer';
 import { Table } from 'reactstrap';
+import { useFindTimeDiff } from '../../hooks/useFindTimeDiff';
 
 const MyComponentToPrint = forwardRef((props, ref) => {
-
-    const findTimes = (start, end) => {
-        let d1 = Date.parse(`2023-01-15T${start}:00.000`);
-        let d2 = Date.parse(`2023-01-15T${end}:00.000`);
-        const milliseconds = Math.abs(d1 - d2);
-        const secs = Math.floor(milliseconds / 1000);
-        const mins = Math.floor(secs / 60);
-        const minutes = mins % 60
-        const hours = Math.floor(mins / 60);
-        return {
-            mins: mins,
-            hours: hours,
-            minutes: minutes,
-            combined: `${hours}hr. ${minutes}min.`
-        }
-    }
+    const { findTimes } = useFindTimeDiff()
 
     let helpersParsedData = []
     if (props.value.helperTimes) {
