+++ src/hooks/useFindTimeDiff.js
@@ -1,0 +1,21 @@
+import { useState } from "react"
+
+export const useFindTimeDiff = () => {
+
+    const findTimes = (start, end) => {
+        let d1 = Date.parse(`2023-01-15T${start}:00.000`);
+        let d2 = Date.parse(`2023-01-15T${end}:00.000`);
+        const milliseconds = Math.abs(d1 - d2);
+        const secs = Math.floor(milliseconds / 1000);
+        const mins = Math.floor(secs / 60);
+        const minutes = mins % 60
+        const hours = Math.floor(mins / 60);
+        return {
+            mins: mins,
+            hours: hours,
+            minutes: minutes,
+            combined: `${hours}hr. ${minutes}min.`
+        }
+    }
+    return { findTimes }
+}
